+++ includes/AnyCommentCore.php
@@ -23,7 +23,7 @@ class AnyCommentCore extends BaseObject {
 	/**
 	 * @var string AnyComment version.
 	 */
-	public $version = '0.1.27';
+	public $version = '0.1.30';
 
 	/**
 	 * @var Pool
@@ -218,22 +218,11 @@ class AnyCommentCore extends BaseObject {
 				@file_put_contents( $index_path, '' );
 			}
 
-			$date_format = 'Y-m-d';
-			$file_format = '{date}-debug';
-			$log_file    = $log_path . DIRECTORY_SEPARATOR;
-			$log_file    .= str_replace(
-				                '{date}',
-				                date( $date_format ),
-				                $file_format
-			                ) . '.log';
-
-			// create a log channel
 			$log     = new Logger( 'anycomment' );
 			$level   = ANYCOMMENT_DEBUG ? Logger::DEBUG : Logger::INFO;
-			$rotator = new RotatingFileHandler( $log_file, 3, $level, true, 0664 );
-			$rotator->setFilenameFormat( $file_format, $date_format );
-			$log->pushHandler( $rotator );
-			$log->pushHandler( new StreamHandler( $log_file, $level ) );
+			$handler = new RotatingFileHandler( $log_path . DIRECTORY_SEPARATOR . 'debug.log', 3, $level, true, 0664 );
+			$handler->setFilenameFormat( '{date}-debug', 'Y-m-d' );
+			$log->pushHandler( $handler );
 
 			static::$log = $log;
 		}